Cambios minimos
Correcciones y validaciones mínimas

+++ tienda/app/Http/Controllers/AdminController.php
@@ -45,24 +45,33 @@ class AdminController extends Controller
     }
 
     public function loadImages(Request $request){
-        $validateData = $request->validate([
-            'file' => 'required|image|mimes:jpeg,jpg,png,gif|max:10000',
-            'file1' => 'required|image|mimes:jpeg,jpg,png,gif|max:10000',
-        ]);
-        $file = $request->file('file');
-        $file1 = $request->file('file1');
-        $finalPath = 'images/homepage/';
-        $fileName = time().'-'.$file->getClientOriginalName();
-        $file1Name = time().'-'.$file1->getClientOriginalName();
-        File::deleteDirectory(public_path('images/homepage/'));
-        $uploadSuccess=$request->file('file')->move($finalPath,$fileName);
-        $uploadSuccess=$request->file('file1')->move($finalPath,$file1Name);
-        DB::table('main_pages')->truncate();
-        $mainPage = new mainPage;
-        $mainPage->file=($finalPath.$fileName);
-        $mainPage->file1=($finalPath.$file1Name);
-        $mainPage->save();
-        return redirect('/home');
+        if($request!=null&&Auth::user){
+            if(Auth::user()->type=='admin'){
+                $validateData = $request->validate([
+                    'file' => 'required|image|mimes:jpeg,jpg,png,gif|max:10000',
+                    'file1' => 'required|image|mimes:jpeg,jpg,png,gif|max:10000',
+                ]);
+                $file = $request->file('file');
+                $file1 = $request->file('file1');
+                $finalPath = 'images/homepage/';
+                $fileName = time().'-'.$file->getClientOriginalName();
+                $file1Name = time().'-'.$file1->getClientOriginalName();
+                File::deleteDirectory(public_path('images/homepage/'));
+                $uploadSuccess=$request->file('file')->move($finalPath,$fileName);
+                $uploadSuccess=$request->file('file1')->move($finalPath,$file1Name);
+                DB::table('main_pages')->truncate();
+                $mainPage = new mainPage;
+                $mainPage->file=($finalPath.$fileName);
+                $mainPage->file1=($finalPath.$file1Name);
+                $mainPage->save();
+                return redirect('/home');
+            }
+            else{
+                return redirect('/home');
+            }
+        }else{
+            return redirect('/home');
+        }
     }
 
     public function upgrade($id){
